feat: reorganize IPC request and response types for conversation agents
