+++ apps/desktop/src/shared/ipc/conversationAgents/conversationAgentAddRequest.ts
