+++ apps/desktop/src/shared/ipc/conversationAgents/conversationAgentRemoveRequest.ts
