+++ apps/desktop/src/shared/ipc/conversationAgents/conversationAgentRemoveResponse.ts
