+++ apps/desktop/src/shared/ipc/index.ts
@@ -42,97 +42,97 @@ export type { SerializableError } from "./types";
 
 // Request types
 export type { SettingsLoadRequest } from "./load";
-export type { SettingsSaveRequest } from "./save";
 export type { SettingsResetRequest } from "./reset";
+export type { SettingsSaveRequest } from "./save";
 
 // Roles request types
 export type { RolesLoadRequest } from "./roles/loadRequest";
-export type { RolesSaveRequest } from "./roles/saveRequest";
 export type { RolesResetRequest } from "./roles/resetRequest";
+export type { RolesSaveRequest } from "./roles/saveRequest";
 
 // Personalities request types
 export type { PersonalitiesLoadRequest } from "./personalities/loadRequest";
-export type { PersonalitiesSaveRequest } from "./personalities/saveRequest";
 export type { PersonalitiesResetRequest } from "./personalities/resetRequest";
+export type { PersonalitiesSaveRequest } from "./personalities/saveRequest";
 
 // Agents request types
 export type { AgentsLoadRequest } from "./agents/loadRequest";
-export type { AgentsSaveRequest } from "./agents/saveRequest";
 export type { AgentsResetRequest } from "./agents/resetRequest";
+export type { AgentsSaveRequest } from "./agents/saveRequest";
 
 // LLM Config request types
 export type { LlmConfigCreateRequest } from "./llmConfig/createRequest";
-export type { LlmConfigReadRequest } from "./llmConfig/readRequest";
-export type { LlmConfigUpdateRequest } from "./llmConfig/updateRequest";
 export type { LlmConfigDeleteRequest } from "./llmConfig/deleteRequest";
-export type { LlmConfigListRequest } from "./llmConfig/listRequest";
 export type { LlmConfigInitializeRequest } from "./llmConfig/initializeRequest";
+export type { LlmConfigListRequest } from "./llmConfig/listRequest";
+export type { LlmConfigReadRequest } from "./llmConfig/readRequest";
 export type { LlmConfigRefreshCacheRequest } from "./llmConfig/refreshCacheRequest";
+export type { LlmConfigUpdateRequest } from "./llmConfig/updateRequest";
 
 // Conversations request types
 export type { ConversationsCreateRequest } from "./conversations/createRequest";
-export type { ConversationsListRequest } from "./conversations/listRequest";
-export type { ConversationsGetRequest } from "./conversations/getRequest";
-export type { ConversationsUpdateRequest } from "./conversations/updateRequest";
 export type { ConversationsDeleteRequest } from "./conversations/deleteRequest";
+export type { ConversationsGetRequest } from "./conversations/getRequest";
+export type { ConversationsListRequest } from "./conversations/listRequest";
+export type { ConversationsUpdateRequest } from "./conversations/updateRequest";
 
 // Conversation Agents request types
-export type { ConversationAgentGetByConversationRequest } from "./requests/conversationAgentGetByConversationRequest";
-export type { ConversationAgentAddRequest } from "./requests/conversationAgentAddRequest";
-export type { ConversationAgentRemoveRequest } from "./requests/conversationAgentRemoveRequest";
-export type { ConversationAgentListRequest } from "./requests/conversationAgentListRequest";
+export type { ConversationAgentAddRequest } from "./conversationAgents/conversationAgentAddRequest";
+export type { ConversationAgentGetByConversationRequest } from "./conversationAgents/conversationAgentGetByConversationRequest";
+export type { ConversationAgentListRequest } from "./conversationAgents/conversationAgentListRequest";
+export type { ConversationAgentRemoveRequest } from "./conversationAgents/conversationAgentRemoveRequest";
 
 // Response types
 export type { SettingsLoadResponse } from "./loadResponse";
-export type { SettingsSaveResponse } from "./saveResponse";
 export type { SettingsResetResponse } from "./resetResponse";
+export type { SettingsSaveResponse } from "./saveResponse";
 export type { SettingsSetDebugLoggingResponse } from "./setDebugLoggingResponse";
 
 // Roles response types
 export type { RolesLoadResponse } from "./roles/loadResponse";
-export type { RolesSaveResponse } from "./roles/saveResponse";
 export type { RolesResetResponse } from "./roles/resetResponse";
+export type { RolesSaveResponse } from "./roles/saveResponse";
 
 // Personalities response types
 export type { PersonalitiesLoadResponse } from "./personalities/loadResponse";
-export type { PersonalitiesSaveResponse } from "./personalities/saveResponse";
 export type { PersonalitiesResetResponse } from "./personalities/resetResponse";
+export type { PersonalitiesSaveResponse } from "./personalities/saveResponse";
 
 // Agents response types
 export type { AgentsLoadResponse } from "./agents/loadResponse";
-export type { AgentsSaveResponse } from "./agents/saveResponse";
 export type { AgentsResetResponse } from "./agents/resetResponse";
+export type { AgentsSaveResponse } from "./agents/saveResponse";
 
 // LLM Config response types
 export type { LlmConfigCreateResponse } from "./llmConfig/createResponse";
-export type { LlmConfigReadResponse } from "./llmConfig/readResponse";
-export type { LlmConfigUpdateResponse } from "./llmConfig/updateResponse";
 export type { LlmConfigDeleteResponse } from "./llmConfig/deleteResponse";
-export type { LlmConfigListResponse } from "./llmConfig/listResponse";
 export type { LlmConfigInitializeResponse } from "./llmConfig/initializeResponse";
+export type { LlmConfigListResponse } from "./llmConfig/listResponse";
+export type { LlmConfigReadResponse } from "./llmConfig/readResponse";
 export type { LlmConfigRefreshCacheResponse } from "./llmConfig/refreshCacheResponse";
+export type { LlmConfigUpdateResponse } from "./llmConfig/updateResponse";
 
 // Conversations response types
 export type { ConversationsCreateResponse } from "./conversations/createResponse";
-export type { ConversationsListResponse } from "./conversations/listResponse";
-export type { ConversationsGetResponse } from "./conversations/getResponse";
-export type { ConversationsUpdateResponse } from "./conversations/updateResponse";
 export type { ConversationsDeleteResponse } from "./conversations/deleteResponse";
+export type { ConversationsGetResponse } from "./conversations/getResponse";
+export type { ConversationsListResponse } from "./conversations/listResponse";
+export type { ConversationsUpdateResponse } from "./conversations/updateResponse";
 
 // Conversation Agents response types
-export type { ConversationAgentGetByConversationResponse } from "./responses/conversationAgentGetByConversationResponse";
-export type { ConversationAgentAddResponse } from "./responses/conversationAgentAddResponse";
-export type { ConversationAgentRemoveResponse } from "./responses/conversationAgentRemoveResponse";
-export type { ConversationAgentListResponse } from "./responses/conversationAgentListResponse";
+export type { ConversationAgentAddResponse } from "./conversationAgents/conversationAgentAddResponse";
+export type { ConversationAgentGetByConversationResponse } from "./conversationAgents/conversationAgentGetByConversationResponse";
+export type { ConversationAgentListResponse } from "./conversationAgents/conversationAgentListResponse";
+export type { ConversationAgentRemoveResponse } from "./conversationAgents/conversationAgentRemoveResponse";
 
 // LLM Models request types
 export type { LlmModelsSaveRequest } from "./llmModels/saveRequest";
 
 // LLM Models response types
 export type { LlmModelsLoadResponse } from "./llmModels/loadResponse";
-export type { LlmModelsSaveResponse } from "./llmModels/saveResponse";
 export type { LlmModelsResetResponse } from "./llmModels/resetResponse";
+export type { LlmModelsSaveResponse } from "./llmModels/saveResponse";
 
 // Data types
-export type { PersistedSettingsData } from "./data";
 export type { SettingsCategory } from "@fishbowl-ai/ui-shared";
+export type { PersistedSettingsData } from "./data";
